chore: refactor s3 controller and router, test passing
